+++ src/RoadCaptain/PlannedRoute.cs
@@ -88,6 +88,11 @@ namespace RoadCaptain
             HasCompleted = false;
             SegmentSequenceIndex = 0;
         }
+
+        public override string ToString()
+        {
+            return Name;
+        }
     }
 
     public enum RouteMoveResult