+++ Changelog.md
@@ -15,6 +15,8 @@
 ### Runner
 
 - Open file dialog now re-opens at the last used location instead of always defaulting to My Documents ([issue #27](https://github.com/sandermvanvliet/RoadCaptain/issues/27))
+- You can now select the Rebel Routes from Zwift Insider as pre-built routes ([issue #28](https://github.com/sandermvanvliet/RoadCaptain/issues/28)):
+![Screenshot of runner with Rebel Routes dropdown](./images/runner-rebel-routes.png)
 
 ### Routing